Implement _.reject
This may not be optimal. Tried to use a call to _. filter but
this implementation does all of the "work" of filter and then also
another _.each call when it could instead be implemented without
the _.filter and with only one _.each. Come back to this later
and try to find a new implementation.